+++ src/torchfem/__init__.py
@@ -3,6 +3,7 @@ import warnings
 warnings.filterwarnings("ignore", "Sparse CSR tensor support is in beta")
 
 from .assembly import Assembly, ReferencePoint, ReferencePointHeat  # noqa
+from .axisymmetric import Axisymmetric, AxisymmetricHeat  # noqa
 from .laminate import Laminate  # noqa
 from .planar import Planar, PlanarHeat  # noqa
 from .shell import Shell, ShellHeat  # noqa